Linearization and band function now working
There, however, appears to be one primary issue with the linearization
results... since it depends on a medial-axis defined by points, the
computed slopes are overly imprecise for properly computing an accurate
slope buffer.  A solution to this would be regressing the medial axis in
segments and then using the derivative of that function for defining the
slope-buffer values.